+++ src/main/java/com/zerobase/backend/BackendApplication.java
@@ -2,10 +2,8 @@ package com.zerobase.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
-@EnableJpaAuditing
 public class BackendApplication {
 
 	public static void main(String[] args) {
