+++ src/main/java/com/zerobase/backend/enums/EvaluateBadge.java
@@ -17,4 +17,18 @@ public enum EvaluateBadge {
     this.type = type;
     this.description = description;
   }
+
+  @Getter
+  public enum EvaluateBadgeType {
+
+    POSITIVE("긍정 평가 뱃지"),
+    NEGATIVE("부정 평가 뱃지");
+
+    private final String description;
+
+    EvaluateBadgeType(String description) {
+      this.description = description;
+    }
+  }
+
 }
