+++ src/main/java/com/zerobase/backend/repository/TeamPurchasePaymentRepository.java
@@ -1,7 +1,6 @@
 package com.zerobase.backend.repository;
 
 import com.zerobase.backend.domain.TeamPurchasePayment;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeamPurchasePaymentRepository extends JpaRepository<TeamPurchasePayment, Long> {
