+++ src/main/java/com/zerobase/backend/security/application/AuthApplication.java
@@ -40,8 +40,7 @@ public class AuthApplication {
 
 
   /**
-   * 로그인
-   * 유저, 사업자 공용
+   * 로그인 유저, 사업자 공용
    */
   public SignResponse signin(SignIn request) {
 
@@ -49,21 +48,18 @@ public class AuthApplication {
     String jwtToken = null;
 
     // 해당 이메일이 유저 이메일인지, 사업자 이메일인지 확인
-    boolean isUser = userRepository.existsByEmail(email);
-    boolean isEntrepreneur = entrepreneurRepository.existsByEmail(email);
-
-    // 해당 이메일의 계정 정보가 DB에 없을경우 예외
-    verifyExistAuthentication(isUser, isEntrepreneur);
-
     // 해당 이메일의 사용자가 유저인경우
-    if(isUser) {
+    if (userRepository.existsByEmail(email)) {
       // 로그인 처리 후 jwt token 발행
       jwtToken = signService.userSignIn(request);
     }
     // 해당 이메일의 사용자가 사업자인경우
-    if (isEntrepreneur) {
+    else if (entrepreneurRepository.existsByEmail(email)) {
       // 로그인 처리 후 jwt token 발행
       jwtToken = signService.entrepreneurSignIn(request);
+    } else {
+      // 해당 이메일의 계정 정보가 DB에 없을경우 예외
+      throw new SecurityCustomException(EMAIL_NOT_FOUND);
     }
 
     // 해당 이메일의 refresh token 발행
@@ -97,19 +93,21 @@ public class AuthApplication {
     // 추출한 email로 refresh token 재발급
     String newRefreshToken = refreshTokenService.createRefreshToken(curJwtToken, emailFromRefresh);
 
+    Role role;
     // jwt token 재발급 시작
     // 현재 사용자가 유저인지 사업자인지 확인
-    boolean isUser = userRepository.existsByEmail(emailFromRefresh);
-    boolean isEntrepreneur = entrepreneurRepository.existsByEmail(emailFromRefresh);
-    Role role = null;
-
-    // 해당 이메일의 계정 정보가 DB에 없을경우 예외
-    verifyExistAuthentication(isUser, isEntrepreneur);
-
     // 해당 이메일의 사용자가 유저인경우 role 값을 유저로 할당
-    if (isUser) role = ROLE_USER;
+    if (userRepository.existsByEmail(emailFromRefresh)) {
+      role = ROLE_USER;
+    }
     // 해당 이메일의 사용자가 사업자인경우 role 값을 사업자로 할당
-    if (isEntrepreneur) role = ROLE_ENTREPRENEUR;
+    else if (entrepreneurRepository.existsByEmail(emailFromRefresh)) {
+      role = ROLE_ENTREPRENEUR;
+    }
+    // 해당 이메일의 계정 정보가 DB에 없을경우 예외
+    else {
+      throw new SecurityCustomException(EMAIL_NOT_FOUND);
+    }
 
     // 새로운 jwt token 발행
     String newJwtToken = jwtComponent.createToken(emailFromRefresh, role.name());
@@ -120,13 +118,8 @@ public class AuthApplication {
   private void setAuthenticationToSecurityContext(UserDetails userDetails) {
     Authentication authentication = new UsernamePasswordAuthenticationToken
         (userDetails, "", userDetails.getAuthorities());
-    SecurityContextHolder.getContextHolderStrategy().setContext(new SecurityContextImpl(authentication));
-  }
-
-  private void verifyExistAuthentication(boolean isUser, boolean isEntrepreneur) {
-    if (!isUser && !isEntrepreneur) {
-      throw new SecurityCustomException(EMAIL_NOT_FOUND);
-    }
+    SecurityContextHolder.getContextHolderStrategy()
+        .setContext(new SecurityContextImpl(authentication));
   }
 
   private void verifyMatch(String curJwtToken, String curRefreshToken) {
