+++ src/main/java/com/zerobase/backend/security/exception/SecurityErrorCode.java
@@ -1,6 +1,8 @@
 package com.zerobase.backend.security.exception;
 
-import static org.springframework.http.HttpStatus.*;
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
+import static org.springframework.http.HttpStatus.CONFLICT;
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
